# policyengine_api/openapi_spec.yaml
openapi: 3.0.0
info:
  title: PolicyEngine API
  description: The PolicyEngine API provides programmatic access to PolicyEngine functionalities, including managing policies and retrieving policy-related information.
  version: null
  contact:
    name: PolicyEngine Support
    email: hello@policyengine.org
    url: https://policyengine.org
servers:
  - url: https://api.policyengine.org
    description: Production server
paths:
...
  /{country_id}/household:
    post:
      summary: Set a household's input data
      operationId: post_household
      description: >-
        Set a household's input data for a given country ID. Certified US bundles
        accept spm beside data and store the resolved selection with the household.
        Saved household inputs and SPM selections are immutable. Create a new
        household to change either; household PUT is not supported. Dual-write
        creation preserves the selection in the v2 household content and identity.
        National SPM requires an explicit selection; county SPM reads county_fips
        from household inputs. A state alone does not identify an SPM area.
      parameters:
        - name: country_id
          in: path
          description: The country ID.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [data]
              properties:
                label:
                  type: string
                  nullable: true
                data:
                  type: object
                spm:
                  $ref: '#/components/schemas/SPMSelection'
      responses:
        400:
          $ref: '#/components/responses/SPMValidationError'
        201:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                    nullable: true
                  result:
                    type: object
                    properties:
                      household_id:
                        type: integer
        404:
          description: Invalid country ID.
          content:
            text/html:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  /{country_id}/household/{household_id}:
    get:
      summary: Get a household's input data with a given ID
      operationId: get_household
      description: Get a household's input data with a given country ID and household ID.
      parameters:
        - name: country_id
          in: path
          description: The country ID.
          required: true
          schema:
            type: string
        - name: household_id
          in: path
          description: The household ID.
          required: true
          schema:
            type: string
      responses:
        200:
          description: The household data.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                    nullable: true
                  result:
                    type: object
                    properties:
                      id:
                        type: integer
                      country_id:
                        type: string
                      household_json:
                        type: object
                      household_hash:
                        type: string
                      spm:
                        $ref: '#/components/schemas/SPMSelection'
                      label:
                        type: string
                        nullable: true
                      api_version:
                        type: string
        404:
          description: Household not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        404:
          description: Invalid country ID.
          content:
            text/html:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  /{country_id}/household/{household_id}/policy/{policy_id}:
    get:
      summary: Get a household's output data under a given policy
      operationId: get_household_under_policy
      description: >-
        Get a household's output data under a given policy. Calculations replay
        the household's stored SPM selection and return its actual provenance.
        SPM geography/composition errors are structured 400 responses.
      parameters:
        - name: country_id
          in: path
          description: The country ID.
          required: true
          schema:
            type: string
        - name: household_id
          in: path
          description: The household ID.
          required: true
          schema:
            type: string
        - name: policy_id
          in: path
          description: The policy ID.
          required: true
          schema:
            type: string
      responses:
        400:
          $ref: '#/components/responses/SPMValidationError'
        200:
          description: The household data under a given policy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                    nullable: true
                  result:
                    type: object
                  spm_config:
                    $ref: '#/components/schemas/SPMSelection'
                  spm_provenance:
                    $ref: '#/components/schemas/SPMProvenance'
        404:
          description: Household or policy not found.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        500:
          description: Error calculating household under policy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        404:
          description: Invalid country ID.
          content:
            text/html:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  /{country_id}/calculate:
    post: &household_calculation_operation
      summary: Calculate household and policy without storing data
      operationId: get_calculate
      description: >-
        Calculate requested household variables without storing data. Certified
        US bundles accept spm beside household and policy. National SPM requires
        an explicit selection; county SPM uses household county_fips inputs.
        Geography/composition errors occur only when an SPM dependency is needed,
        so a state-only tax calculation can succeed. Metro selection requires
        geography_id; national and county selections forbid geography_id. The
        selected artifact and scenario must belong to the certified bundle.
      parameters:
        - name: country_id
          in: path
          description: The country ID.
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                household:
                  type: object
                policy:
                  type: object
                spm:
                  $ref: '#/components/schemas/SPMSelection'
      responses:
        400:
          $ref: '#/components/responses/SPMValidationError'
        200:
          description: The calculation result.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
                    nullable: true
                  result:
                    type: object
                  spm_config:
                    $ref: '#/components/schemas/SPMSelection'
                  spm_provenance:
                    $ref: '#/components/schemas/SPMProvenance'
        500:
          description: Error calculating household under policy.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
        404:
          description: Invalid country ID.
          content:
            text/html:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  message:
                    type: string
  /{country_id}/calculate-full:
    post:
      <<: *household_calculation_operation
      operationId: get_calculate_full
      summary: Calculate all household variables without storing data
      description: >-
        Calculate household variables after adding omitted yearly variables.
        Accepts the same household, policy, and certified US spm settings as
        calculate. SPM dependencies require household county_fips, an explicit
        metro area, or an explicit national selection; state alone is insufficient.
...
components:
  schemas:
...
    SPMSelection:
      description: Generated from policyengine_api.spm.SPMSelection.
    SPMProvenance:
      description: Generated from policyengine_api.spm.SPMProvenance.
  responses:
    SPMValidationError:
      description: >-
        Household input or settings validation failed. SPM errors always include
        result null and an errors array with code/message objects; other validation
        errors may omit those fields. SPM household input codes
        are SPM_GEOGRAPHY_REQUIRED, SPM_GEOGRAPHY_UNAVAILABLE,
        SPM_COMPOSITION_REQUIRED, and SPM_YEAR_UNAVAILABLE. Unsupported or uncertified settings use
        SPM_SETTINGS_INVALID, SPM_SETTINGS_UNSUPPORTED, or
        SPM_CONFIGURATION_UNAVAILABLE.
      content:
        application/json:
          schema:
            type: object
            required: [status, message]
            properties:
              status:
                type: string
                enum: [error]
              message:
                type: string
              result:
                type: object
                nullable: true
              errors:
                type: array
                items:
                  type: object
                  required: [message]
                  properties:
                    code:
                      type: string
                    message:
                      type: string